Create Quiz Show: The Numbers Round.py
On the quiz show "Letters and Numbers" there is a round in which contestants are given six ingredient numbers and one target number, and their challenge is to apply arithmetic operations (addition, subtraction, multiplication, division) to one or more of the ingredient numbers to get the target number.
The ingredient numbers are chosen randomly as follows: 0-4 large numbers, chosen from {25, 50, 75, 100} with no repeats and the remaining are all small numbers, chosen from {1, 2, 3, ..., 10} with no repeats.
Each ingredient number can only be used once, but they need not all be used. No fractions are allowed at any stage of the calculation.
The target number is also chosen randomly from 100 to 1000.